refactor: 가격의 타입을 Int 에서 Double로 변경
- 게시물의 가격이 Double형태로 들어있을 때 조회가 안되는 오류 발생
- 가격의 타입을 Int 에서 Double로 변경해주어 해결

+++ OpenMarket/OpenMarket/Model/ModificationData.swift
@@ -12,8 +12,8 @@ struct ModificationData {
     var name: String? = nil
     var descriptions: String? = nil
     var thumbnailId: String? = nil
-    var price: Int? = nil
+    var price: Double? = nil
     var currency: Currency? = nil
-    var discountedPrice: Int? = nil
+    var discountedPrice: Double? = nil
     var stock: Int? = nil
 }
